backend/requirements.txt
fastapi>=0.115.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
sqlalchemy[asyncio]>=2.0.36,<3.0.0
asyncpg>=0.30.0,<1.0.0
pgvector>=0.3.0,<1.0.0
greenlet>=3.0.0,<4.0.0
openai>=1.59.0,<2.0.0
pypdf>=5.0.0,<6.0.0
PyJWT[crypto]>=2.9.0,<3.0.0
